Tell a hosted ACP session its approval policy is degraded (#738)
Every degradation must be user-visible, and the daemon log carrying the same
fact is not. The note goes out on the transcript envelope lane, before the
opening turn, so it stays ordered with the session and lands ahead of anything
the agent does under the weakened policy.
